patching apl lib2raw

+++ src/aflowlib_libraries.cpp
@@ -7554,13 +7554,15 @@ namespace aflowlib {
       //ME20211008 - Check for AFLOW modules
       //APL
       apl_aflowin_found = apl::APL_Get_AflowInName(AflowInName, directory_LIB);
-      aurostd::RemoveFile(directory_LIB+"/"+POCC_FILE_PREFIX + POCC_APL_OUT_FILE);
-      FileLockName = "LOCK.apl";
-      if(aurostd::EFileExist(directory_LIB+"/"+FileLockName, stmp) && aurostd::IsCompressed(stmp)) {
-        aurostd::UncompressFile(stmp);
+      if(apl_aflowin_found){  //CO20211125
+        aurostd::RemoveFile(directory_LIB+"/"+POCC_FILE_PREFIX + POCC_APL_OUT_FILE);
+        FileLockName = "LOCK.apl";
+        if(aurostd::EFileExist(directory_LIB+"/"+FileLockName, stmp) && aurostd::IsCompressed(stmp)) {
+          aurostd::UncompressFile(stmp);
+        }
+        vAflowInName.push_back(AflowInName);
+        vFileLockName.push_back(FileLockName);
       }
-      vAflowInName.push_back(AflowInName);
-      vFileLockName.push_back(FileLockName);
     } else {
       // [OBSOLETE] else if(aurostd::FileExist(directory_LIB+"/agl_aflow.in"))
       AGL_functions::AGL_Get_AflowInName(AflowInName, directory_LIB, agl_aflowin_found); //CT20200713 Call function to find correct aflow.in file name  //CO20210204 - fix aflow.in.xz inside
